% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manuals.R
\docType{data}
\name{Distance_data_master}
\alias{Distance_data_master}
\title{Locations data with geo spatial distance}
\usage{
data(Distance_data_master)
}
\description{
A data frame consisting of all available locations along with spatial disatnce between all combinations
}
