% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocations.R
\name{getAllLocations}
\alias{getAllLocations}
\title{Get All locations of NOAA Weather Stations}
\usage{
getAllLocations(online = TRUE, state = NA)
}
\arguments{
\item{online}{if TRUE location data is pulled from the API, if FALSE, data is pulled from offiline database}

\item{state}{filter the final output based on the state code specified, deafault to all US states}
}
\value{
Dataframe with Location ID, Location name and State
}
\description{
get all locations information using API call
}
\examples{
locations <- getAllLocations (online=FALSE,state="CA")
}
