% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetrics.R
\name{getMetrics}
\alias{getMetrics}
\title{get all Avaialble Metrics}
\usage{
getMetrics(online = TRUE)
}
\arguments{
\item{online}{if TRUE metrics data is pulled from the API, if FALSE, data is pulled from offiline database}
}
\value{
All metrics Information
}
\description{
get all Avaialble Metrics from NOAA Web API along with descriptions and units
}
