% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpacialPlotData.R
\name{getSpatialPlotData}
\alias{getSpatialPlotData}
\title{generate data for plot_spatial()}
\usage{
getSpatialPlotData(online = TRUE, date = "2017-01-01",
  metric = "t_official")
}
\arguments{
\item{online}{if TRUE(default),realtime data is pulled from the API. if FALSE, data is pulled from sample offline database}

\item{date}{date for which spatial plot is needed. Default: 2017-01-01}

\item{metric}{metric for which spatial plot is needed. Default: "t_official"}
}
\value{
a dataframe of values for the given metric and date
}
\description{
generate data for plot_spatial()
}
