% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manuals.R
\docType{data}
\name{Metrics}
\alias{Metrics}
\title{Metrics available via NOAA Weather API}
\usage{
data(Metrics)
}
\description{
A data frame consisting of all metrics (with definations, units) details
}
