% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calhmap.R
\name{plot_calhmap}
\alias{plot_calhmap}
\title{calendar heat map of weather data}
\usage{
plot_calhmap(online = TRUE, from = "2017-01-01", to = "2017-01-31",
  measure = "t_max", location = "Wolf Point", state = "MT", text = TRUE)
}
\arguments{
\item{online}{if TRUE(default),realtime data is pulled from the API. if FALSE, data is pulled from sample offline database}

\item{from}{Beginning of date range. Default: 2017-01-01}

\item{to}{End of date range. Default: 2017-01-31}

\item{measure}{metric for which time series plot is needed. Default: "t_max"}

\item{location}{Valid location supported by NOAA, use \code{\link{getAllLocations}} for view avaialble locations. Default: "Wolf Point"}

\item{state}{Valid two letter code of US state. Default: "MT"}

\item{text}{if TRUE, print the values of each day on heatmap, if FALSE values are not printed}
}
\value{
calendar heat map for a given time period, measure, location, and state
}
\description{
generate calendar heat map for a given time period, measure, location, and state
}
\examples{
plot_calhmap(online=FALSE,from="2017-01-01",
to="2017-01-03", measure="t_max",
location='Austin', state='TX')
}
