% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manuals.R
\docType{data}
\name{weather_data}
\alias{weather_data}
\title{Weather data for 2017 January}
\usage{
data(weather_data)
}
\description{
A data frame consisting of hourly data for metrics: ("p_official","rh_std","solarad","t_max","t_min","t_official","windspd","ws_max")
}
